Close the overflow hole the raw-only guard did not reach
The re-review found that widening the span comparison in ServesFullResolution
was not enough on its own. ComputeBucket carries the same unchecked 64 bit
subtraction, so a request from long.MinValue to long.MaxValue wrapped the range
to -1, produced no bucket, and the timeline handler took the unaggregated branch
regardless of the new guard. Confirmed against the endpoint: it returned every
row in the table.

Both halves are needed, because the handler aggregates only when the exemption
is lost AND a bucket exists. Widen ComputeBucket the same way. The widest bucket
this can produce is about 9.2e15, so the result still fits a long, and for any
range that did not overflow the arithmetic is unchanged, so no ordinary request
shifts resolution.

This is pre-existing code the story did not otherwise touch, and it is fixed
here rather than filed because issue #29 asks for the timeline response to be
bounded, and it was not yet bounded for every input the endpoint accepts.

The existing overflow test now asserts the bucket as well as the exemption, and
a new end-to-end test requests the full range of long over HTTP. Verified that
both fail without the ComputeBucket change.

+++ viewer/TierSelection.cs
@@ -164,12 +164,18 @@ public static class TierSelection
 
     static long ComputeBucket(long from, long to, List<TierSlice> slices)
     {
-        long rangeMs = to - from;
+        // Widened for the same reason as TierPlan.ServesFullResolution: from and
+        // to reach here unvalidated, and a range spanning most of long overflows
+        // a 64 bit subtraction into a negative number. That reads as a range too
+        // narrow to bucket, which would leave the widest request expressible as
+        // the one that comes back unaggregated. The widest bucket this can
+        // produce is about 9.2e15, so the result still fits a long.
+        Int128 rangeMs = (Int128)to - from;
         long coarsest = slices.Max(s => NativeIntervalMs(s.Table));
         long bucket = 0;
 
         if (rangeMs > 0 && rangeMs / coarsest > MaxPoints)
-            bucket = (rangeMs / MaxPoints / coarsest) * coarsest;
+            bucket = (long)(rangeMs / MaxPoints / coarsest) * coarsest;
 
         // Mixed tiers must share one bucket size, otherwise the series would
         // change resolution partway along the chart.
